gaiakeyutil -> gaiacli keys parse
Incorporate gaiakeyutil functionalities into gaiacli.
